crm: convert limits menu
